requirements.txt
requests==2.32.3
psycopg[binary]==3.2.3
psycopg_pool==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.1
